+++ src/Http/routes.php
@@ -2,10 +2,10 @@
 
 Route::group(['middleware' => 'web'], function () {
 
-    Route::group(['prefix' => 'admin/pages', 'middleware' => ['auth'], 'namespace' => 'Ourgarage\StaticPages\Http\Controllers'], function () {
+    Route::group(['prefix' => 'admin/pages', 'middleware' => ['auth'], 'namespace' => 'Ourgarage\StaticPages\Http\Controllers\Admin'], function () {
 
-        Route::get('/', 'StaticPagesController@indexAdmin')->name('static-pages::admin::index');
-        Route::get('/create', 'StaticPagesController@createPage')->name('static-pages::admin::create-page');
+        Route::get('/', 'StaticPagesController@index')->name('static-pages::admin::index');
+        Route::get('/create', 'StaticPagesController@create')->name('static-pages::admin::create-page');
         Route::post('/status-update/{id}', 'StaticPagesController@statusUpdate')->name('static-pages::admin::status-update');
         Route::delete('/delete/{id}', 'StaticPagesController@destroy')->name('static-pages::admin::page-delete');
         Route::post('/store/{id?}', 'StaticPagesController@store')->name('static-pages::admin::page-store');
