+++ src/resources/views/site/pages-list.blade.php
