+++ src/resources/views/site/pages-view.blade.php
